+++ CHANGELOG.rst
@@ -13,12 +13,13 @@ Current changes
 
 Added
 -----
-- Add background image and SSID/PIN information display.
 - Add flake8 test.
 - Add more tests.
 
 Changed
 -------
+- Show UI with gtk3.
+- Use gstreamer for player.
 - Use oxmplayer for rtp player.
 - wfd_video_formats: update native-resolution.